src/storyboard_video/infra/fonts.py
from PIL import ImageFont


def load_font(size: int) -> ImageFont.FreeTypeFont | ImageFont.ImageFont:
    for candidate in ["msyh.ttc", "msyhbd.ttc", "simhei.ttf", "arial.ttf"]:
        try:
            return ImageFont.truetype(candidate, size)
        except Exception:
            continue
    return ImageFont.load_default()
